Print out serialized translated term.  Add a makefile.
If Ocaml is installed, the translate bytecode executable can be built
with "make" or "make translate".

dart2ast.dart takes a filename and writes a serialized Dart AST to
stdout.  translate reads a serialized Dart AST from stdin, translates
it, and serialized the result to stdout.  The serialized output is not
yet pretty printed.  So to run on test.dart:

./dart2ast.dart test.dart | ./translate