Add support for 'dispose' to sync* iterators.
In the Single and Nested cases which correspond to yield and yield*,
add a thunk which can be called to run all finally blocks in the sync*
body.

This requires passing a finally continuation to the translation
function for statements, but not expressions (because yield and yield*
are statements in Dart).

The translation of Yield (yield), YieldStar (yield*), and
YieldBreak (return from a sync* body) are changed so that they do not
invoke the return continuation of the translation.  This continuation
represents returning from the translated function itself and not the
return from the inner function used to construct the returned
iterable.