Do not name argument subexpressions unless needed.
Argument subexpressions only need to be named if there is an await in
an argument subexpression to their right.  To achieve this,
responsibility for naming argument subexpressions is moved from when
visiting the expression itself to when visiting its parent, which
knows if there is an await to the right.

Then to avoid naming values (identifiers, constants) they must be
inspected before adding a binding.