Add statements to the analysis pass and support IfStatement.
Statements are analyzed and marked if the have a subexpression
containing await.  This is used to avoid transforming statements that do
not contain any await.

Support is added for IfStatement.  If there is an await in either the
then or else part, then a local join-point function is used to avoid
duplicating continuations.  This can be improved in at least two ways:

1. The join-point function is only needed if there is an unconditional
   await in one of the bodies.

2. The join-point function should be trampolined.