Add support for break/continue from while loops.
The translation keeps a stack of break targets and a stack of continue
targets.  These are Dart expressions representing unary functions that
ignore their argument (the argument is necessary because they are
invoked by finally blocks).

Labels are not yet supported, so there is no way to test breaking or
continuing to an outer loop.