Fix a bug in try/finally; add support for constants.
In try/finally: (a) there should be no finally block in the
translation output, because all paths already include a call to the
finally code and we do not want the finally code on the 'await' path.
And (b) the body of the try should emit a call to the finally code.

Also, add support for all literals except list, map, and string
interpolation.