Handle failure of 'on T catch' clauses properly.
'on T catch' clauses are translated into if/else with explicit type
tests.  If all the tests fail the exception was thrown again, losing
the stack trace from the original exception.

Instead, use an explicit application of the error continuation (which
we already know there's an await).  This will either be a direct call
to Completer.completeError with two arguments, or a call to a local
catch handler function with two arguments.